+++ grakn-test/test-distribution/src/test/java/ai/grakn/distribution/DistributionE2E.java
@@ -13,11 +13,11 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.concurrent.TimeoutException;
 
-import static ai.grakn.distribution.DistributionITConstants.GRAKN_UNZIPPED_DIRECTORY;
-import static ai.grakn.distribution.DistributionITConstants.assertGraknRunning;
-import static ai.grakn.distribution.DistributionITConstants.assertGraknStopped;
-import static ai.grakn.distribution.DistributionITConstants.assertZipExists;
-import static ai.grakn.distribution.DistributionITConstants.unzipGrakn;
+import static ai.grakn.distribution.DistributionE2EConstants.GRAKN_UNZIPPED_DIRECTORY;
+import static ai.grakn.distribution.DistributionE2EConstants.assertGraknRunning;
+import static ai.grakn.distribution.DistributionE2EConstants.assertGraknStopped;
+import static ai.grakn.distribution.DistributionE2EConstants.assertZipExists;
+import static ai.grakn.distribution.DistributionE2EConstants.unzipGrakn;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
